Make pgoutput plugin the default
The "pgoutput" plugin is the server side default plugin for logical
replication and the one that emits the [Logical Replication Protocol Messages](https://www.postgresql.org/docs/current/protocol-logical-replication.html#PROTOCOL-LOGICAL-MESSAGES).

While a stream of XLogData values might be interesting for passing on or
working with the data transmitted from other plugins, this is probably
the most useful format and what most people want.