Convert timestamp and timestamp with time zone directly to DateTime
There was a very old ToDo-comment in the code because TimestampHandler
and TimestampTzHandler were passing values through NpgsqlDateTime in
order to convert them from/to the backend representation.
For Replication we need to convert timestamps in a completely separate
part of the codebase so we change it now and benefit from it multiple
times.