Make ReplicationConnection.PostgreSqlVersion public + add ServerVersion
Now that we have two versions of the Logical Streaming Replication
Protocol people may be interested in the backend version they are
connected to in order to adjust the requested protocol version.
Publicly exposing ReplicationConnection.PostgreSqlVersion will help
them to identify the backend version without having to open a separate
NpgsqlConnection.